Move error related utils to separate file
