+++ sdk/servicebus/service-bus/src/session/messageSession.ts
@@ -26,11 +26,8 @@ import {
 } from "../core/messageReceiver";
 import { LinkEntity } from "../core/linkEntity";
 import { ClientEntityContext } from "../clientEntityContext";
-import {
-  convertTicksToDate,
-  calculateRenewAfterDuration,
-  throwErrorIfConnectionClosed
-} from "../util/utils";
+import { convertTicksToDate, calculateRenewAfterDuration } from "../util/utils";
+import { throwErrorIfConnectionClosed } from "../util/errors";
 import { ServiceBusMessage, DispositionType, ReceiveMode } from "../serviceBusMessage";
 import { messageDispositionTimeout } from "../util/constants";
 
@@ -322,7 +319,7 @@ export class MessageSession extends LinkEntity {
         this._onError(error);
         log.error(
           "[%s] Notified the user's error handler about the error received by the " +
-          "Receiver '%s'.",
+            "Receiver '%s'.",
           this._context.namespace.connectionId,
           this.name
         );
@@ -338,7 +335,7 @@ export class MessageSession extends LinkEntity {
           this._context.expiredMessageSessions[this.sessionId!] = true;
           sbError.message = `The session lock has expired on the session with id ${
             this.sessionId
-            }.`;
+          }.`;
         }
         log.error(
           "[%s] An error occurred for Receiver '%s': %O.",
@@ -377,7 +374,7 @@ export class MessageSession extends LinkEntity {
         }
         log.error(
           "[%s] 'receiver_close' event occurred for receiver '%s' for sessionId '%s'. " +
-          "The associated error is: %O",
+            "The associated error is: %O",
           connectionId,
           this.name,
           this.sessionId,
@@ -389,7 +386,7 @@ export class MessageSession extends LinkEntity {
       if (receiver && !receiver.isItselfClosed()) {
         log.error(
           "[%s] 'receiver_close' event occurred on the receiver '%s' for sessionId '%s' " +
-          "and the sdk did not initiate this. Hence, let's gracefully close the receiver.",
+            "and the sdk did not initiate this. Hence, let's gracefully close the receiver.",
           connectionId,
           this.name,
           this.sessionId
@@ -408,7 +405,7 @@ export class MessageSession extends LinkEntity {
       } else {
         log.error(
           "[%s] 'receiver_close' event occurred on the receiver '%s' for sessionId '%s' " +
-          "because the sdk initiated it. Hence no need to gracefully close the receiver",
+            "because the sdk initiated it. Hence no need to gracefully close the receiver",
           connectionId,
           this.name,
           this.sessionId
@@ -428,7 +425,7 @@ export class MessageSession extends LinkEntity {
         const sbError = translate(sessionError);
         log.error(
           "[%s] 'session_close' event occurred for receiver '%s' for sessionId '%s'. " +
-          "The associated error is: %O",
+            "The associated error is: %O",
           connectionId,
           this.name,
           this.sessionId,
@@ -441,7 +438,7 @@ export class MessageSession extends LinkEntity {
       if (receiver && !receiver.isSessionItselfClosed()) {
         log.error(
           "[%s] 'session_close' event occurred on the receiver '%s' for sessionId '%s' " +
-          "and the sdk did not initiate this. Hence, let's gracefully close the receiver.",
+            "and the sdk did not initiate this. Hence, let's gracefully close the receiver.",
           connectionId,
           this.name,
           this.sessionId
@@ -460,7 +457,7 @@ export class MessageSession extends LinkEntity {
       } else {
         log.error(
           "[%s] 'session_close' event occurred on the receiver '%s' for sessionId '%s' " +
-          "because the sdk initiated it. Hence no need to gracefully close the receiver",
+            "because the sdk initiated it. Hence no need to gracefully close the receiver",
           connectionId,
           this.name,
           this.sessionId
@@ -486,7 +483,7 @@ export class MessageSession extends LinkEntity {
       if (this._sessionLockRenewalTimer) clearTimeout(this._sessionLockRenewalTimer);
       log.messageSession(
         "[%s] Cleared the timers for 'no new message received' task and " +
-        "'session lock renewal' task.",
+          "'session lock renewal' task.",
         this._context.namespace.connectionId
       );
       if (this._receiver) {
@@ -559,7 +556,7 @@ export class MessageSession extends LinkEntity {
           const msg =
             `MessageSession '${this.sessionId}' with name '${this.name}' did not receive ` +
             `any messages in the last ${
-            this.newMessageWaitTimeoutInSeconds
+              this.newMessageWaitTimeoutInSeconds
             } seconds. Hence closing it.`;
           log.error("[%s] %s", this._context.namespace.connectionId, msg);
 
@@ -587,7 +584,7 @@ export class MessageSession extends LinkEntity {
         ) {
           log.error(
             "[%s] Not calling the user's message handler for the current message " +
-            "as the receiver '%s' is closed",
+              "as the receiver '%s' is closed",
             connectionId,
             this.name
           );
@@ -608,7 +605,7 @@ export class MessageSession extends LinkEntity {
           if (!isAmqpError(err)) {
             log.error(
               "[%s] An error occurred while running user's message handler for the message " +
-              "with id '%s' on the receiver '%s': %O",
+                "with id '%s' on the receiver '%s': %O",
               connectionId,
               bMessage.messageId,
               this.name,
@@ -627,7 +624,7 @@ export class MessageSession extends LinkEntity {
             try {
               log.error(
                 "[%s] Abandoning the message with id '%s' on the receiver '%s' since " +
-                "an error occured: %O.",
+                  "an error occured: %O.",
                 connectionId,
                 bMessage.messageId,
                 this.name,
@@ -638,7 +635,7 @@ export class MessageSession extends LinkEntity {
               const translatedError = translate(abandonError);
               log.error(
                 "[%s] An error occurred while abandoning the message with id '%s' on the " +
-                "receiver '%s': %O.",
+                  "receiver '%s': %O.",
                 connectionId,
                 bMessage.messageId,
                 this.name,
@@ -673,7 +670,7 @@ export class MessageSession extends LinkEntity {
             const translatedError = translate(completeError);
             log.error(
               "[%s] An error occurred while completing the message with id '%s' on the " +
-              "receiver '%s': %O.",
+                "receiver '%s': %O.",
               connectionId,
               bMessage.messageId,
               this.name,
@@ -810,7 +807,7 @@ export class MessageSession extends LinkEntity {
             const msg =
               `MessageSession '${this.sessionId}' with name '${this.name}' did not receive ` +
               `any messages in the last ${
-              this.newMessageWaitTimeoutInSeconds
+                this.newMessageWaitTimeoutInSeconds
               } seconds. Hence closing it.`;
             log.error("[%s] %s", this._context.namespace.connectionId, msg);
             finalAction();
@@ -935,7 +932,7 @@ export class MessageSession extends LinkEntity {
         this._deliveryDispositionMap.delete(delivery.id);
         log.receiver(
           "[%s] Disposition for delivery id: %d, did not complete in %d milliseconds. " +
-          "Hence resolving the promise.",
+            "Hence resolving the promise.",
           this._context.namespace.connectionId,
           delivery.id,
           messageDispositionTimeout
@@ -990,7 +987,7 @@ export class MessageSession extends LinkEntity {
       if (!this.isOpen() && !this.isConnecting) {
         log.error(
           "[%s] The receiver '%s' with address '%s' is not open and is not currently " +
-          "establishing itself. Hence let's try to connect.",
+            "establishing itself. Hence let's try to connect.",
           connectionId,
           this.name,
           this.address
@@ -1068,7 +1065,7 @@ export class MessageSession extends LinkEntity {
       } else {
         log.error(
           "[%s] The receiver '%s' for sessionId '%s' is open -> %s and is connecting " +
-          "-> %s. Hence not reconnecting.",
+            "-> %s. Hence not reconnecting.",
           connectionId,
           this.name,
           this.sessionId,
@@ -1138,7 +1135,7 @@ export class MessageSession extends LinkEntity {
         try {
           log.messageSession(
             "[%s] Attempting to renew the session lock for MessageSession '%s' " +
-            "with name '%s'.",
+              "with name '%s'.",
             connectionId,
             this.sessionId,
             this.name
@@ -1153,7 +1150,7 @@ export class MessageSession extends LinkEntity {
           );
           log.receiver(
             "[%s] Successfully renewed the session lock for MessageSession '%s' " +
-            "with name '%s'.",
+              "with name '%s'.",
             connectionId,
             this.sessionId,
             this.name
@@ -1167,7 +1164,7 @@ export class MessageSession extends LinkEntity {
         } catch (err) {
           log.error(
             "[%s] An error occurred while renewing the session lock for MessageSession " +
-            "'%s' with name '%s': %O",
+              "'%s' with name '%s': %O",
             this._context.namespace.connectionId,
             this.sessionId,
             this.name,
@@ -1177,7 +1174,7 @@ export class MessageSession extends LinkEntity {
       }, nextRenewalTimeout);
       log.messageSession(
         "[%s] MessageSession '%s' with name '%s', has next session lock renewal " +
-        "in %d seconds @(%s).",
+          "in %d seconds @(%s).",
         this._context.namespace.connectionId,
         this.sessionId,
         this.name,
