+++ sdk/servicebus/service-bus/src/subscriptionClient.ts
@@ -12,7 +12,7 @@ import {
   getOpenReceiverErrorMsg,
   throwErrorIfClientOrConnectionClosed,
   throwErrorIfConnectionClosed
-} from "./util/utils";
+} from "./util/errors";
 import { AmqpError, generate_uuid } from "rhea-promise";
 import { ClientEntityContext } from "./clientEntityContext";
