+++ sdk/servicebus/service-bus/src/topicClient.ts
@@ -9,7 +9,7 @@ import {
   getOpenSenderErrorMsg,
   throwErrorIfClientOrConnectionClosed,
   throwErrorIfConnectionClosed
-} from "./util/utils";
+} from "./util/errors";
 import { AmqpError, generate_uuid } from "rhea-promise";
 import { ClientEntityContext } from "./clientEntityContext";
